完成下载部分，摒弃mvc封装的download，采用三方PRDownloader实现。 完善GameInfActivity(部分功能不再实现具体仅供查看)

+++ findgame/src/main/java/com/example/findgame/recommend/controller/HttpUtil.java
@@ -1,6 +1,23 @@
 package com.example.findgame.recommend.controller;
 
+import android.content.Context;
+import android.os.Handler;
+import android.os.Message;
+import android.text.StaticLayout;
+import android.text.method.SingleLineTransformationMethod;
 import android.util.Log;
+import android.widget.Toast;
+
+import com.downloader.Error;
+import com.downloader.OnCancelListener;
+import com.downloader.OnDownloadListener;
+import com.downloader.OnPauseListener;
+import com.downloader.OnProgressListener;
+import com.downloader.OnStartOrResumeListener;
+import com.downloader.PRDownloader;
+import com.downloader.PRDownloaderConfig;
+import com.downloader.Progress;
+import com.example.findgame.bean.GameInfBean;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -23,6 +40,9 @@ import okhttp3.RequestBody;
 import okhttp3.Response;
 import okhttp3.ResponseBody;
 import okhttp3.logging.HttpLoggingInterceptor;
+import okio.Buffer;
+import okio.BufferedSink;
+import okio.Sink;
 
 public class HttpUtil {
 
@@ -31,7 +51,7 @@ public class HttpUtil {
     private OkHttpClient okHttpClient;
 
 
-    private HttpUtil(){
+    private HttpUtil() {
 
         //okhttp消息拦截器
         HttpLoggingInterceptor httpLoggingInterceptor = new HttpLoggingInterceptor();
@@ -46,28 +66,29 @@ public class HttpUtil {
             }
         };
 
-        okHttpClient=new OkHttpClient.Builder()
+        okHttpClient = new OkHttpClient.Builder()
                 .readTimeout(1, TimeUnit.MINUTES)
-                .connectTimeout(1,TimeUnit.MINUTES)
+                .connectTimeout(1, TimeUnit.MINUTES)
                 .addInterceptor(interceptor)
                 .addInterceptor(httpLoggingInterceptor)
                 .build();
 
     }
 
-    private static HttpUtil httpUtil=null;
-    public static HttpUtil getInstance(){
-        if (httpUtil == null){
-            synchronized (Object.class){
-                if (httpUtil == null){
-                    httpUtil=new HttpUtil();
+    private static HttpUtil httpUtil = null;
+
+    public static HttpUtil getInstance() {
+        if (httpUtil == null) {
+            synchronized (Object.class) {
+                if (httpUtil == null) {
+                    httpUtil = new HttpUtil();
                 }
             }
         }
         return httpUtil;
     }
 
-    public void doGet(String url, final OKutil oKutil){
+    public void doGet(String url, final OKutil oKutil) {
         Log.i(TAG, "doGet: ");
         final Request request = new Request.Builder()
                 .url(url)
@@ -90,14 +111,14 @@ public class HttpUtil {
         });
     }
 
-    public  void doPost(String url, HashMap<String,String> map, OKutil listener){
+    public void doPost(String url, HashMap<String, String> map, OKutil listener) {
         FormBody.Builder builder = new FormBody.Builder();
-        Set<Map.Entry<String, String>> entrise=map.entrySet();
+        Set<Map.Entry<String, String>> entrise = map.entrySet();
 
         for (Map.Entry<String, String> entry : entrise) {
             String key = entry.getKey();
             String value = entry.getValue();
-            builder.add(key,value);
+            builder.add(key, value);
         }
         FormBody body = builder.build();
 
@@ -119,13 +140,13 @@ public class HttpUtil {
             }
         });
     }
-    public void doDownload(String url, final String path, final MvcListener linstener){
+
+    public void doDownload(String url, final String path, final MvcListener linstener) {
 
         Request builder = new Request.Builder()
                 .url(url)
                 .get()
                 .build();
-
         Call call = okHttpClient.newCall(builder);
         call.enqueue(new Callback() {
             @Override
@@ -136,19 +157,19 @@ public class HttpUtil {
             @Override
             public void onResponse(Call call, Response response) throws IOException {
                 ResponseBody body = response.body();
-                long max=body.contentLength();
+                long max = body.contentLength();
 
                 InputStream inputStream = body.byteStream();
                 FileOutputStream fileOutputStream = new FileOutputStream(path);
                 byte[] bytes = new byte[1024];
-                int count=0;
-                int len=0;
-                while ((len=inputStream.read(bytes))!=-1){
-                    count+=len;
-                    fileOutputStream.write(bytes,0,len);
-                    linstener.onProgress((int) (count*100/max));
+                long count = 0;
+                int len = 0;
+                while ((len = inputStream.read(bytes)) != -1) {
+                    count += len;
+                    fileOutputStream.write(bytes, 0, len);
+                    linstener.onProgress((int) (count * 100 / max));
                 }
-                if (count>=max){
+                if (count >= max) {
                     linstener.onFinish();
                 }
             }
@@ -156,10 +177,10 @@ public class HttpUtil {
     }
 
 
-    public void doUpload(String url, String path, String filename, String type, final MvcListener linstener){
+    public void doUpload(String url, String path, String filename, String type, final MvcListener linstener) {
         MultipartBody builder = new MultipartBody.Builder()
                 .setType(MultipartBody.FORM)
-                .addFormDataPart("file",filename, RequestBody.create(MediaType.parse(type),new File(path)))
+                .addFormDataPart("file", filename, RequestBody.create(MediaType.parse(type), new File(path)))
                 .build();
 
         Request request = new Request.Builder().url(url).post(builder).build();
